Use socket instead of USB
gbridge application added ability to greybus to connect
to a module through TCP/IP.
USB was required to emulate an APBridge but that's not
required anymore because we can use TCP/IP as transport layer.
So, completly remove USB.

This patch alone is not enough to support Greybus over TCP/IP.
In addition of emulating an APBridge, gbsim is also acting like
a SVC, so we have also to get rid of SVC code.

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>