Load the manifest blob at boot
Currently, the manifest is load by inotify,
by putting the blob in hotplug folder.
This give the ability to gbsim to change the manifest at runtime.
But it is quite hard to handle it with the change made to support TCP/IP.
Update gbsim to load it once at boot.

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>